+++ plugin/hdCycles/renderBuffer.cpp
@@ -286,7 +286,7 @@ HdCyclesRenderBuffer::BlitTile(HdFormat format, unsigned int x, unsigned int y,
 
         for (unsigned int j = 0; j < height; ++j) {
             for (unsigned int i = 0; i < width; ++i) {
-                int mem_start = (((y + j) * m_width) * m_pixelSize)
+                size_t mem_start = (((y + j) * m_width) * m_pixelSize)
                                 + ((x + i) * m_pixelSize);
 
                 int tile_mem_start = ((j * width) * pixelSize)
